lib: move the sg13g2 P&R data into ihp130 and drop the companion modules
ihp130_pnr.py and ihp130_stdcells.py are deleted. pdk() now lists the
stdcell LEF, GDS and SPICE paths, and a design builds its own ExtLibrary
from them. sg13g2_grid() and lef_pin_rects(), which parses the LEF itself
now, move into ihp130.py. The engine's is_leaf parameter defaults to the
new is_extlibrary_leaf, an external-library cell is a routing leaf, so
is_sg13g2_leaf is gone and designs no longer pass a predicate.
