+++ src/ordec/layout/digital_pnr/__init__.py
@@ -6,7 +6,7 @@ A gridded standard-cell place-and-route engine: cells into abutted rows, then
 all signal nets routed on a track grid by negotiated congestion.
 
 The engine is PDK-agnostic. Every pitch, layer and DRC dimension arrives as an
-explicit parameter of :func:`place_and_route`, which :mod:`ordec.lib.ihp130_pnr`
+explicit parameter of :func:`place_and_route`, which :mod:`ordec.lib.ihp130`
 supplies for the IHP sg13g2 standard cells.
 """
