+++ tests/test_digital_pnr.py
@@ -24,10 +24,9 @@ from ordec.layout.digital_pnr import (GridConfig, PinAccessError,
     place_and_route)
 from ordec.layout.digital_pnr import place, route
 from ordec.layout.digital_pnr.flow import (LeafCell, NetInfo,
-    flatten_schematic, stack_from_spec)
+    flatten_schematic, is_extlibrary_leaf, stack_from_spec)
 from ordec.layout.digital_pnr.route import M2, M3, VERT
-from ordec.lib.ihp130 import SG13G2
-from ordec.lib.ihp130_pnr import is_sg13g2_leaf, lef_pin_rects, sg13g2_grid
+from ordec.lib.ihp130 import SG13G2, lef_pin_rects, sg13g2_grid
 from .lib import pnr_cells as fx
 
 # A grid with the sg13g2 dimensions but no PDK behind it, so the placement and
@@ -51,7 +50,7 @@ def pnr(cell, layout=None, port_edges=None):
         layout = Layout(cell=cell, symbol=cell.symbol)
     result = place_and_route(cell.schematic, layout, grid=sg13g2_grid(),
         routing_spec=SG13G2().default_routing_spec, pin_rects=lef_pin_rects,
-        is_leaf=is_sg13g2_leaf, port_edges=port_edges)
+        port_edges=port_edges)
     return layout.freeze(), result
 
 
@@ -259,7 +258,7 @@ def test_flatten_expands_composites():
     """A composite sub-cell is replaced by its own leaves, with its internal
     nets uniquified by the instance prefix and its ports rewired."""
     leaves, nets = flatten_schematic(fx.PairChain(n=3).schematic,
-        is_sg13g2_leaf)
+        is_extlibrary_leaf)
 
     # Three InvPairs of two inverters each, none of them composite any more.
     assert sorted(leaves) == [f"pr[{i}]/i{k}" for i in range(3) for k in (0, 1)]
